Move definition of allFacetFields into a helper file
In order to avoid deoendency cycle between search result and facet
browser components.
