+++ src/components/facet-browser/helper.ts
@@ -1,0 +1,16 @@
+import { FacetField } from "../../core/dbc-gateway/generated/graphql";
+
+export const allFacetFields = [
+  FacetField.MainLanguages,
+  FacetField.AccessTypes,
+  FacetField.ChildrenOrAdults,
+  FacetField.Creators,
+  FacetField.FictionNonfiction,
+  FacetField.FictionalCharacter,
+  FacetField.GenreAndForm,
+  FacetField.MaterialTypes,
+  FacetField.Subjects,
+  FacetField.WorkTypes
+];
+
+export default {};
